Deleted and moved embed utility classes
